Split up test directories by es version.
There are some small differences in the code required to test the
different ES versions supported by JanusGraph. The code to support http
in the ElasticSearchIndex, however, is the same. Therefore, I've only
split up the test directories. If at some point in the future it's
required to split up that file based off of version, we can do so then.
