+++ janusgraph-es-parent/janusgraph-es-1x-test/.gitignore
@@ -1,0 +1,2 @@
+/bin
+/config
