+++ janusgraph-es-parent/janusgraph-es-1x-test/src/test/config/networkNodeUsingYaml.yml
