+++ janusgraph-es-parent/janusgraph-es-1x-test/src/test/java/org/janusgraph/diskstorage/es/ElasticsearchStatus.java
