+++ janusgraph-es-parent/janusgraph-es-1x-test/src/test/resources/es_cfg_bogus_nodeclient.yml
