+++ janusgraph-es-parent/janusgraph-es-1x-test/src/test/resources/es_nodename_bar.yml
