+++ janusgraph-es-parent/janusgraph-es-1x-test/src/test/resources/log4j.properties
