+++ janusgraph-es-parent/janusgraph-es-test/pom.xml
@@ -1,0 +1,130 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>org.janusgraph</groupId>
+        <artifactId>janusgraph-es-parent</artifactId>
+        <version>0.1.0-SNAPSHOT</version>
+        <relativePath>../pom.xml</relativePath>
+    </parent>
+    <artifactId>janusgraph-es-test</artifactId>
+    <dependencies>
+        <dependency>
+            <groupId>org.janusgraph</groupId>
+            <artifactId>janusgraph-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.janusgraph</groupId>
+            <artifactId>janusgraph-test</artifactId>
+            <version>${project.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.janusgraph</groupId>
+            <artifactId>janusgraph-es-core-test</artifactId>
+            <version>${project.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.janusgraph</groupId>
+            <artifactId>janusgraph-berkeleyje</artifactId>
+            <version>${project.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.janusgraph</groupId>
+            <artifactId>janusgraph-berkeleyje</artifactId>
+            <version>${project.version}</version>
+            <classifier>tests</classifier>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.janusgraph</groupId>
+            <artifactId>janusgraph-cassandra</artifactId>
+            <version>${project.version}</version>
+            <classifier>tests</classifier>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.janusgraph</groupId>
+            <artifactId>janusgraph-cassandra</artifactId>
+            <version>${project.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.antlr</groupId>
+            <artifactId>antlr-runtime</artifactId>
+            <version>${antlr.version}</version>
+        </dependency>
+        <!--Bit of a hack here. Jest relies on a later version of gson. This
+          should be fine though. -->
+        <dependency>
+            <groupId>com.google.code.gson</groupId>
+            <artifactId>gson</artifactId>
+            <version>2.2.4</version>
+        </dependency>
+        <dependency>
+          <groupId>io.searchbox</groupId>
+          <artifactId>jest</artifactId>
+          <version>2.0.4</version><!-- change this to appropriate jar for ES 1.x -->
+          <exclusions>
+            <exclusion>  <!-- declare the exclusion here -->
+              <groupId>com.google.code.gson</groupId>
+              <artifactId>gson</artifactId>
+            </exclusion>
+          </exclusions>
+        </dependency>
+    </dependencies>
+    <profiles>
+        <profile>
+            <id>es-1</id>
+            <activation>
+                 <property>
+                    <name>elasticsearch.version</name>
+                    <value>[1.0, 2.0)</value>
+                 </property>
+            </activation>
+            <dependencies>
+                <dependency>
+                  <groupId>org.apache.lucene</groupId>
+                  <artifactId>lucene-test-framework</artifactId>
+                  <!--TODO move this to max version variable -->
+                  <version>5.5.2</version>
+                  <scope>test</scope>
+                </dependency>
+                <dependency>
+                  <groupId>org.elasticsearch</groupId>
+                  <artifactId>elasticsearch</artifactId>
+                  <!--TODO move this to max version variable -->
+                  <version>2.4.4</version>
+                  <scope>test</scope>
+                  <type>test-jar</type>
+                </dependency>
+            </dependencies>
+        </profile>
+        <profile>
+            <id>es-2+</id>
+            <activation>
+                 <property>
+                    <name>elasticsearch.version</name>
+                    <value>[2.0,)</value>
+                 </property>
+            </activation>
+            <dependencies>
+                <dependency>
+                  <groupId>org.apache.lucene</groupId>
+                  <artifactId>lucene-test-framework</artifactId>
+                  <version>${lucene.version}</version>
+                  <scope>test</scope>
+                </dependency>
+                <dependency>
+                  <groupId>org.elasticsearch</groupId>
+                  <artifactId>elasticsearch</artifactId>
+                  <version>${elasticsearch.version}</version>
+                  <scope>test</scope>
+                  <type>test-jar</type>
+                </dependency>
+            </dependencies>
+        </profile>
+    </profiles>
+</project>
