+++ janusgraph-es-parent/janusgraph-es/bin/elasticsearch.in.sh
