+++ janusgraph-es-parent/janusgraph-es/config/logging.yml
