+++ janusgraph-es-parent/janusgraph-es/src/main/resources/janusgraph-es.properties
