+++ janusgraph-es-parent/janusgraph-es/src/test/binassembly.xml
@@ -1,0 +1,17 @@
+<assembly xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+          xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2 http://maven.apache.org/xsd/assembly-1.1.2.xsd">
+    <id>es-bin</id>
+    <baseDirectory>/</baseDirectory>
+
+    <formats>
+        <format>dir</format>
+    </formats>
+    
+    <fileSets>
+        <fileSet>
+            <directory>src/test/bin</directory>
+            <outputDirectory>/</outputDirectory>
+            <filtered>false</filtered>
+        </fileSet>
+    </fileSets>
+</assembly>
