+++ janusgraph-es-parent/janusgraph-es/src/test/resources/es_cfg_bogus_nodeclient.yml
@@ -1,0 +1,5 @@
+node.data: false
+node.client: true
+discovery.zen.ping.multicast.enabled: false
+discovery.zen.ping.unicast.hosts: [ "10.11.12.13" ]
+cluster.name: networkNodeUsingYaml
