+++ janusgraph-es-parent/janusgraph-es/src/test/resources/es_cfg_nodeclient.yml
@@ -1,0 +1,5 @@
+node.data: false
+node.client: true
+discovery.zen.ping.multicast.enabled: false
+discovery.zen.ping.unicast.hosts: [ "localhost", "127.0.0.1:9300" ]
+cluster.name: networkNodeUsingYaml
