+++ janusgraph-es-parent/janusgraph-es/src/test/resources/es_nodename_bar.yml
@@ -1,0 +1,1 @@
+node.name: bar
